analyze/1 now also takes a single filename
percept2:analyze/1 expects a list of filenames even if you only
provide a single one. When you provide a single filename instead
of a list, you will get errors like:

  {error,"Incorrect file or malformed trace file: 47\n"}

Those are not easily identifiable as wrong input parameters.

With this commit it is now possible to supply a single filename
that is is not wrapped in a list to percept2:analyze/1.